+++ src/autora/state/history.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, Iterable, List, Optional, Sequence, Set, Union
 from numpy.typing import ArrayLike
 from sklearn.base import BaseEstimator
 
-from autora.state.delta import Delta
+from autora.state import Delta
 from autora.state.protocol import (
     ResultKind,
     SupportsControllerStateHistory,
