+++ src/autora/state/wrapper.py
@@ -11,7 +11,7 @@ from typing import Callable, TypeVar
 import pandas as pd
 from sklearn.base import BaseEstimator
 
-from autora.state.delta import Delta, State, on_state
+from autora.state import Delta, State, on_state
 from autora.variable import VariableCollection
 
 S = TypeVar("S")
